move from py2 to 3
works perfectly for yara and filename iocs, not a byte difference in the logs on 1600 hits. but I didn't have test cases for wmi and process scanning.

it's not backwards compatible to python2 because it was awkward at some places with the conversions between old unicode, bytes, str, ... and py2 is end of life anyway.